Remove OpenG dependency
The `Register Check Running.vi` uses an OpenG VI to filter an array
based on if the array was empty or not. The usage of this OpenG VI
required the installation of the OpenG Array library/package, which is
not a big deal, except for two reasons.

First, the OpenG VI dependency is only used within the `Register Check
Running.vi` within the Caraya project. It feels a little excessive to
add just a dependency for a single usage, especially given the relative
simplicity of the functionality.

Second, the OpenG Array library is installed into `user.lib` instead of
`vi.lib` for legacy reasons and this causes problems when using Caraya
in a project and attempting to save for previous version. When using the
Save for Previous Version... action from the __File>>Save for Previous
Version...__ of the _Project Explorer_, the `vi.lib` is excluded, but
the `user.lib` is not. The OpenG dependency then gets saved along with
the rest of the project with the full folder hierarchy from the C drive
on a Windows machine be included. For example, LabVIEW and the OpenG
dependency are installed in the default locations on a Windows machine,
i.e. `C:\Program Files (x86)\National Instruments\LabVIEW 2017` and
`C:\Program Files (x86)\National Instruments\LabVIEW 2017\user.lib\_OpenG.lib\array\array.lib`, and I have a project saved at `C:\Users\example_user\Documents\Projects\Example Using Caraya`. When I save for previous version the Example Using Caraya project to an output folder at `C:\Users\example_user\Documents\Projects\output`, I will get the following folder structure in `output`:

output
|  C
|  +  Program Files (x86)
|  |  +  National Instruments
|  |  |  +  LabVIEW 2017
|  |  |  |  +  user.lib
|  |  |  |  |  +  _OpenG.lib
|  |  |  |  |  |  +  array
|  |  |  |  |  |  |  +  array.lib
|  |  |  |  |  |  |  |  ...
|  +  Users
|  |  +  example_user
|  |  |  +  Documents
|  |  |  |  +  Projects
|  |  |  |  |  +  Example Using Caraya
|  |  |  |  |  |  ...

This is quite cumbersome and excessive to deal with when attempting to
create a package for distribution, especially when only a single VI from
the dependency is only used in one location within the Caraya project.

Ideally, the solution would be to have the `user.lib` folder ignored
from the save for previous version action, but ignoring other folders,
such as `instr.lib` or `user.lib`, appears to be under evaluation by NI
(https://forums.ni.com/t5/LabVIEW/Excluding-instr-lib-folder-from-Project-Save-For-Previous/td-p/3760288),
but no timeline is given. Another solution is to have the OpenG
dependency installed in `vi.lib` as opposed to `user.lib`, but this is
most likely impossible because of other projects, add-ons, and packages
depending on OpenG being installed in `user.lib`.

Thus, I have come up with a workaround. I have eliminated the need for
the OpenG array dependency by simply replacing its usage with a custom
implementation. Now, the `user.lib` does not get included in the save
for previous version action and the number of dependencies is reduced.