Fix Rerun All Tests opens About dialog when there are no test cases
For an empty project, the test manager currently opens the first VI
it finds in the application directory when the user selects Rerun All
Tests. This happens because there is no root test (empty string).

Incidentally, this results in About.vi being launched because it is
the first VI in a sorted list of files, which is not what the user
expects.

Fixed by checking if the root test is empty, in which case no action
is taken (silently fails).

References #139