Fix Run Tests in Active Project opens two dialogs
VIs called from the tools menu execute in a different context than
ones called from a project. Because of this, FGVs don't share their
state between contexts, which results in two dialogs being created.
One for the context of the tools menu (NI.LV.Dialog) and one for the
project (My Computer).

Fixed by launching the test runner in the context of the project.

Closes #138